requirements.txt
absl-py==2.1.0
addict==2.4.0
basicsr==1.4.2
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
facexlib==0.2.5
filelock==3.16.1
filterpy==1.4.5
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
gfpgan==1.3.5
grpcio==1.67.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
realesrgan==0.3.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
tb-nightly==2.19.0a20241102
tensorboard-data-server==0.7.2
tifffile==2024.9.20
tokenizers==0.20.1
tomli==2.0.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
transformers==4.46.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.3
Werkzeug==3.1.1
yapf==0.40.2
zipp==3.20.2
